Refactor `update_evs.js`
This got split into 2 Stimulus controllers, `ev` and `goals`.
